# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Docs",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/v1/notebook": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get all notes",
                "operationId": "d85f51a68364b1841236171170d37ae5",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "full_name": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            },
                                            "phone_number": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "date_birth": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "path_to_photo": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Create a new note",
                "operationId": "8f118254f72ba6c05d4ccdbdeabaf18c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone_number",
                                    "email"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Example Name"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Company"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example.name@example.com"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-12-26"
                                    },
                                    "path_to_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note has been created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Note has been created"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "full_name": {
                                                    "type": "string"
                                                },
                                                "company": {
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "date_birth": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "path_to_photo": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get a note by ID",
                "operationId": "32a71159918fa7b65f744a1f0d36c7e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "full_name": {
                                            "type": "string"
                                        },
                                        "company": {
                                            "type": "string"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "date_birth": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "path_to_photo": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update a note",
                "operationId": "92ccfacdef6c3cff5139e976e57e8ee5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone_number",
                                    "email"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Example Name"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Company"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example.name@example.com"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-11-26"
                                    },
                                    "path_to_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note has been updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Note has been updated"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "full_name": {
                                                    "type": "string"
                                                },
                                                "company": {
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "date_birth": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "path_to_photo": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete a note",
                "operationId": "ebed6dd65da75aee313eb25c5f792d2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note has been deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            }
        }
    }
}